% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundleArchive.R
\name{bundleArchive}
\alias{bundleArchive}
\title{Bundles the model archive into a .tar archive to share with reviewers.}
\usage{
bundleArchive(filePath = "modelArchive", bundleName = "modelArchive.tar.gz")
}
\arguments{
\item{filePath}{Relative file path of folder cooresponding to model archive folder.}

\item{bundleName}{The filename of the tar archive that holds the model archive bundle.}
}
\description{
The \code{bundleArchive} function Bundles the model archive into a .tar archive to share with reviewers.
}
\examples{
bundleArchive()
}

