% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundlePackages.R
\name{bundlePackages}
\alias{bundlePackages}
\title{Bundles all packages required by model}
\usage{
bundlePackages(filePath = "modelArchive", repositories = c(CRAN =
  "http://cran.rstudio.com", GRAN = "http://owi.usgs.gov/R"), localRepos = "")
}
\arguments{
\item{filePath}{Relative file path of folder cooresponding to model archive folder.}

\item{repositories}{Named character vector containing urls of package repositories}

\item{localRepos}{Character string containing path to local local repository on disk for packages, if applicable.}
}
\description{
The \code{bundlePackages} function obtains all packages other than packages included in base R used in the model and bundles them for inclusion in the model archive.
}
\examples{
bundlePackages()
}

