% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkArchive.R
\name{checkArchive}
\alias{checkArchive}
\title{Checks basic files and documentation in model archive.}
\usage{
checkArchive(filePath = "modelArchive")
}
\arguments{
\item{filePath}{Relative file path of folder cooresponding to model archive folder.}
}
\description{
The \code{checkArchive} function checks that all files are documented in the readme and that each folder contains files.
}
\examples{
makeReadMe()
}

