% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyScripts.R
\name{copyScripts}
\alias{copyScripts}
\title{Copies script files to scripts folder in model archive}
\usage{
copyScripts(scriptFiles = NULL, filePath = "modelArchive/scripts")
}
\arguments{
\item{scriptFiles}{character vector of script file names to copy to scripts directory in model archive folder. If NULL defaults to all files in working directory with .R extension.}

\item{filePath}{file path of folder to copy scripts to.}
}
\description{
The \code{copyScripts} function downloads the appropriate version of R to include in the archive.
}
\examples{
copyScripts()
}

