% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFolder.R
\name{createFolder}
\alias{createFolder}
\title{Create a model archive folder}
\usage{
createFolder(folderPath = NULL, folderName = "modelArchive")
}
\arguments{
\item{folderPath}{full file path to directory to place model archive folder. If left NULL places folder in current working directory.}

\item{folderName}{character string of name for model archive folder.}
}
\description{
The \code{createFolder} function creates a model archive folder with sub directories for required files. Defaults to current working directory.
}
\examples{
createFolder()
}

