% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getR.R
\name{getR}
\alias{getR}
\title{Downloads R version to include with archive}
\usage{
getR(OS = "win", rVersion = NULL, filePath = "modelArchive/R_files")
}
\arguments{
\item{OS}{Operating system for R version. Can be one of "win","mac","linux", or any combination of the three.}

\item{rVersion}{Version of R to include, e.g. "3.1.1". Defaults to current version of R in use.}

\item{filePath}{Relative file path to store downloaded R files.}
}
\description{
The \code{getR} function downloads the appropriate version of R to include in the archive.
}
\examples{
createFolder()
}

