% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unbundleArchive.R
\name{unbundleArchive}
\alias{unbundleArchive}
\title{Opens a bundled archive.}
\usage{
unbundleArchive(bundle, where)
}
\arguments{
\item{bundle}{The full path to the bundled model archive.}

\item{where}{The full path to where the unbundled archive is to be placed.}
}
\description{
The \code{unbundleArchive} function opens a model archive created using \code{bundleArchive} and sets the current R library to that stored in the bundled archive.
}
\examples{
unbundleArchive(bundle="sharedArchive.tar.gz",where = "myReviewFolder")
}

