% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\name{dataDoc}
\alias{dataDoc}
\title{Checks basic files and documentation in model archive.}
\usage{
dataDoc(dataFrame, file = NULL)
}
\arguments{
\item{dataFrame}{Data.frame object to document}

\item{file}{Filename of text file to contain output, should be the same name as datafile being described with no extension. If left NULL output prints to console.}
}
\description{
The \code{checkArchive} function checks that all files are documented in the readme and that each folder contains files.
}
\examples{
dataDoc(dataFrame = iris)
}

